server/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic>=2.0.0
supabase==2.*
python-dotenv==1.*
httpx==0.27.2
beautifulsoup4==4.12.3
openai>=1.0.0
